+++ src/utility/ostream_bit_writer.cpp
@@ -150,7 +150,7 @@ void ostream_bit_writer::write_variable_bits_big_endian(uint64_t value,
 {
     // TODO: should be an exception
     auto remaining_bits = std::min(static_cast<uint64_t>(least_significant_bits),
-        (byte_bits * sizeof(uint64_t)));
+        static_cast<uint64_t>(byte_bits * sizeof(uint64_t)));
 
     while (remaining_bits > byte_bits)
     {
